refactor(services): extract JobLineage from JobStore (#451 step 4/5) (#502)
Wave 6.4, fourth of the JobStore decomposition. Pull the Re-tune/Resume
lineage graph + per-parent retune lock (H-0062) out of `JobStore` into
its own module.

New module `services/_job_lineage.py` (200 lines) — `JobLineage`:
- `__init__(metadata: JobMetadataStore)` — holds the in-memory
  `_parent_locks: dict[parent_id, child_id]` + `_parent_lock_mutex`;
  delegates disk access to the injected metadata store (`jobs_dir` for
  the directory scan, `get` for full-job loads)
- `get_child_job_ids` / `get_lineage_tree` / `has_active_children` —
  the parent→child queries (derived from each job's `parent_job_id`),
  moved verbatim
- `acquire_parent_lock` / `release_parent_lock` / `rebind_parent_lock` /
  `get_locked_child` — the per-parent exclusive lock, moved verbatim

`JobStore.__init__` constructs `self._lineage = JobLineage(self._meta)`;
drops `_parent_locks` / `_parent_lock_mutex`; the seven public lineage
methods are one-line delegations. `delete`'s cascade BFS keeps calling
`self.get_child_job_ids(...)` (now the delegation). `json` import
dropped from `jobs.py` (only used in the moved `get_child_job_ids`).

The parent-lock tests (`test_parent_lock.py`, `test_retune_api.py`,
`test_training_service.py` retune-lock cases, `test_reg_0077_cross_parent_retune_race.py`)
and lineage tests pass unchanged — they assert through the public
`acquire_parent_lock` / `get_locked_child` / `get_lineage_tree` / ...
surface.

Pure refactor, no public API or behaviour change — Change Gate not
required. `jobs.py` 622 → 511 lines. No test changes needed.

Verified: full non-slow backend suite (1533 passed, unchanged), ruff /
mypy clean.

Refs #451 (step 4/5). Next: 5/5 — JobStore becomes a thin orchestrator
(target < 200 lines): trim the `delete` cascade + `reconcile_at_startup`
+ `set_status` to use the collaborators directly, slim the module to the
wiring + the orchestration that genuinely spans concerns.
